pyAgreeZone/requirements.txt
amqp==2.5.2
asgiref==3.2.3
billiard==3.6.1.0
celery==4.4.0
certifi==2019.11.28
chardet==3.0.4
cos-python-sdk-v5==1.7.7
dicttoxml==1.7.4
Django==2.1.8
django-filter==2.2.0
django-redis==4.11.0
djangorestframework==3.11.0
idna==2.8
importlib-metadata==1.4.0
kombu==4.6.7
more-itertools==8.1.0
python-crontab==2.4.0
python-dateutil==2.8.1
pytz==2019.3
qcloud-python-sts==3.0.3
redis==3.3.11
requests==2.22.0
six==1.14.0
sqlparse==0.3.0
tencentcloud-sdk-python==3.0.121
urllib3==1.25.7
vine==1.3.0
zipp==1.0.0
